Add "legacy" commands for serial & DFU mode management w/ baud-switcher
Adds appropriate wrapper fns for serial open/close and dfu open/close
as subcommands for the "legacy" command.  Also adds docs / help strings.

tested and working on linux

only thing that may need to still be added is an update to "completion.py"
to account for the new fns but it was unclear if the overall format would
need to be changed to support doing so, so this was left off this commit